Remove deprecated prisma migrate files
